Fixed the name of useProfileForm.tsx
Modified useProfileForm hook

Split the code for better organization
