+++ frontend/src/components/settings/form/SettingsForm.tsx
@@ -4,7 +4,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { Form } from "@/components/ui/form";
 import { ProfileFormType, profileFormSchema } from "@/lib/types/profile-form";
-import useProfileFrom from "@/hooks/usProfileForm";
+import useProfileFrom from "@/hooks/useProfileForm";
 import { cn } from "@/lib/utils";
 import SettingsFormImageFields from "./SettingsFormImageFields";
 import SettingsFormFields from "./SettingsFormFields";
